fix: table pagination behavior
